+++ backend/src/app.module.ts
@@ -3,7 +3,7 @@ import { ConfigModule, ConfigService } from "@nestjs/config";
 import { MongooseModule } from "@nestjs/mongoose";
 import { AuthModule } from './auth/auth.module';
 import { StudentModule } from './student/student.module';
-import { TeacherModule } from './teacher/infrastruture/teacher.module';
+import { TeacherModule } from './teacher/teacher.module'; 
 import { AdminModule } from './admin/admin.module'; 
 import { ParentModule } from './parent/parent.module'; 
 import { v2 as cloudinary } from 'cloudinary';
