--- backend/src/teacher/domain/teacher.interface.ts
@@ -1,16 +1,0 @@
-
-
-
-export interface Teacher {
-  firstName: string;
-  lastName: string;
-  email: string;
-  mobileNumber: string;
-  profileImage:string;
-  dob: Date;
-  degree?: string;
-  experienceYears: number;
-  subjects?: string[];
-  address:string;
-  univerity:string;
-}
